v12

+++ get_next_line.c
@@ -172,7 +172,7 @@ int get_next_line(int const fd, char **line)
 	}
 	while ((buffsize = read(fd, &buff, BUFF_SIZE)) > 0)
 	{
-		if (buffsize == -1)
+		if (buffsize == (size_t)-1)
 			return (-1);
 		buff[buffsize] = '\0';
 		if ((addstk = ft_add_stack(&stack, buff)) == -1)
